Changed libav decoder in <analyseH264_libav.c> to use libavformat to read input stream, rather than passing the raw file. Seems to have fixed the video artifacts. Bugfixes to the HWM handling in cameraControl, to stop it thinking that products of rawvideos are ahead of the HWM due to having timestamps up to an hour later.

+++ src/cameraControl/daytimeJobsClean.py
@@ -29,6 +29,7 @@ for taskGroup in dayTimeTasks:
       for dirName, subdirList, fileList in os.walk(outDir):
         for f in fileList:
           utc = module_hwm.filenameToUTC(f)
+          if (utc < 0): continue
           if (utc > highWaterMarks[HWMout]): os.system("rm -f %s"%os.path.join(dirName,f))
 
 os.chdir(CWD)
